Удалена лишняя функция

+++ score.cpp
@@ -1,7 +1,7 @@
 #include "score.hpp"
 
-Score::Score(double _mg, double _eg) : mg(_mg), eg(_eg) {}
+Score::Score(int _mg, int _eg) : mg(_mg), eg(_eg) {}
 
-double Score::getScore(double stageGame) const {
+int Score::getScore(int stageGame) const {
     return mg * stageGame + (1 - stageGame) * eg;
 }
